+++ core/base/discreteGradient/DiscreteGradient_Template.h
@@ -1066,6 +1066,8 @@ int DiscreteGradient::buildGradient(){
     numberOfCells[i]=getNumberOfCells(i);
 
   dmtMax2PL_.clear();
+  dmt1Saddle2PL_.clear();
+  dmt2Saddle2PL_.clear();
   gradient_.clear();
   gradient_.resize(dimensionality_);
   for(int i=0; i<dimensionality_; ++i){
@@ -1153,8 +1155,14 @@ template <typename dataType>
 int DiscreteGradient::setCriticalPoints(const std::vector<Cell>&
 criticalPoints) const{
   const dataType* const scalars=static_cast<dataType*>(inputScalarField_);
+  const int* const offsets=static_cast<int*>(inputOffsets_);
   std::vector<dataType>* outputCriticalPoints_points_cellScalars=
     static_cast<std::vector<dataType>*>(outputCriticalPoints_points_cellScalars_);
+
+  const auto sosGreaterThan=[&scalars,&offsets](const int a, const int b){
+    if(scalars[a] != scalars[b]) return scalars[a]>scalars[b];
+    else return offsets[a]>offsets[b];
+  };
 
   (*outputCriticalPoints_numberOfPoints_)=0;
 
@@ -1183,17 +1191,56 @@ criticalPoints) const{
     outputCriticalPoints_points_cellIds_->push_back(cellId);
     outputCriticalPoints_points_cellScalars->push_back(scalar);
     outputCriticalPoints_points_isOnBoundary_->push_back(isOnBoundary);
+    int vertexId=-1;
     if(dmtMax2PL_.size()){
       if(cellDim==0)
-        outputCriticalPoints_points_PLVertexIdentifiers_->push_back(cellId);
+        vertexId=cellId;
       else if(cellDim==dimensionality_)
-
-        outputCriticalPoints_points_PLVertexIdentifiers_->push_back(dmtMax2PL_[cellId]);
-      else
-        outputCriticalPoints_points_PLVertexIdentifiers_->push_back(-1);
+        vertexId=dmtMax2PL_[cellId];
     }
-    else
-      outputCriticalPoints_points_PLVertexIdentifiers_->push_back(-1);
+
+    if(dmt1Saddle2PL_.size() and cellDim==1){
+      vertexId=dmt1Saddle2PL_[cellId];
+
+      if(vertexId==-1){
+        simplexId_t v0;
+        simplexId_t v1;
+        inputTriangulation_->getEdgeVertex(cellId, 0, v0);
+        inputTriangulation_->getEdgeVertex(cellId, 1, v1);
+
+        if(sosGreaterThan(v0,v1))
+          vertexId=v0;
+        else
+          vertexId=v1;
+      }
+    }
+    if(dmt2Saddle2PL_.size() and cellDim==2){
+      vertexId=dmt2Saddle2PL_[cellId];
+
+      if(vertexId==-1){
+        simplexId_t v0;
+        simplexId_t v1;
+        simplexId_t v2;
+        if(dimensionality_==2){
+          inputTriangulation_->getCellVertex(cellId, 0, v0);
+          inputTriangulation_->getCellVertex(cellId, 1, v1);
+          inputTriangulation_->getCellVertex(cellId, 2, v2);
+        }
+        else if(dimensionality_==3){
+          inputTriangulation_->getTriangleVertex(cellId, 0, v0);
+          inputTriangulation_->getTriangleVertex(cellId, 1, v1);
+          inputTriangulation_->getTriangleVertex(cellId, 2, v2);
+        }
+        if(sosGreaterThan(v0,v1) and sosGreaterThan(v0,v2))
+          vertexId=v0;
+        else if(sosGreaterThan(v1,v0) and sosGreaterThan(v1,v2))
+          vertexId=v1;
+        else
+          vertexId=v2;
+      }
+    }
+
+    outputCriticalPoints_points_PLVertexIdentifiers_->push_back(vertexId);
 
     (*outputCriticalPoints_numberOfPoints_)++;
   }
@@ -1259,7 +1306,6 @@ int DiscreteGradient::setAugmentedCriticalPoints(const std::vector<Cell>& critic
       if(cellDim==0)
         outputCriticalPoints_points_PLVertexIdentifiers_->push_back(cellId);
       else if(cellDim==dimensionality_)
-
         outputCriticalPoints_points_PLVertexIdentifiers_->push_back(dmtMax2PL_[cellId]);
       else
         outputCriticalPoints_points_PLVertexIdentifiers_->push_back(-1);
@@ -1363,11 +1409,12 @@ template <typename dataType>
 int DiscreteGradient::getRemovableSaddles1(const std::vector<std::pair<simplexId_t,char>>& criticalPoints,
                                            const bool allowBoundary,
                                            std::vector<char>& isRemovableSaddle,
-                                           std::vector<simplexId_t>& pl2dmt_saddle) const{
+                                           std::vector<simplexId_t>& pl2dmt_saddle){
   const simplexId_t numberOfEdges=inputTriangulation_->getNumberOfEdges();
   isRemovableSaddle.resize(numberOfEdges);
 
-  std::vector<char> dmt2PL(numberOfEdges, false);
+  dmt1Saddle2PL_.resize(numberOfEdges);
+  std::fill(dmt1Saddle2PL_.begin(), dmt1Saddle2PL_.end(), -1);
 
   // by default : 1-saddle is removable
 #ifdef TTK_ENABLE_OPENMP
@@ -1396,7 +1443,7 @@ int DiscreteGradient::getRemovableSaddles1(const std::vector<std::pair<simplexId
         inputTriangulation_->getVertexEdge(criticalPointId, i, edgeId);
         const Cell saddleCandidate(1, edgeId);
 
-        if(isSaddle1(saddleCandidate) and !dmt2PL[edgeId]){
+        if(isSaddle1(saddleCandidate) and dmt1Saddle2PL_[edgeId]==-1){
           saddleId=edgeId;
           ++numberOfSaddles;
         }
@@ -1404,8 +1451,8 @@ int DiscreteGradient::getRemovableSaddles1(const std::vector<std::pair<simplexId
 
       // only one DMT-1saddle in the star so this one is non-removable
       if(numberOfSaddles==1){
-        if(!dmt2PL[saddleId] and pl2dmt_saddle[criticalPointId]==-1){
-          dmt2PL[saddleId]=true;
+        if(dmt1Saddle2PL_[saddleId]==-1 and pl2dmt_saddle[criticalPointId]==-1){
+          dmt1Saddle2PL_[saddleId]=criticalPointId;
           pl2dmt_saddle[criticalPointId]=saddleId;
           isRemovableSaddle[saddleId]=false;
         }
@@ -1421,11 +1468,12 @@ int DiscreteGradient::getRemovableSaddles2(const
                                            std::vector<std::pair<simplexId_t,char>>& criticalPoints,
                                            const bool allowBoundary,
                                            std::vector<char>& isRemovableSaddle,
-                                           std::vector<simplexId_t>& pl2dmt_saddle) const{
+                                           std::vector<simplexId_t>& pl2dmt_saddle){
   const simplexId_t numberOfTriangles=inputTriangulation_->getNumberOfTriangles();
   isRemovableSaddle.resize(numberOfTriangles);
 
-  std::vector<char> dmt2PL(numberOfTriangles, false);
+  dmt2Saddle2PL_.resize(numberOfTriangles);
+  std::fill(dmt2Saddle2PL_.begin(), dmt2Saddle2PL_.end(), -1);
 
   // by default : 2-saddle is removable
 #ifdef TTK_ENABLE_OPENMP
@@ -1454,7 +1502,7 @@ int DiscreteGradient::getRemovableSaddles2(const
         inputTriangulation_->getVertexTriangle(criticalPointId, i, triangleId);
         const Cell saddleCandidate(2, triangleId);
 
-        if(isSaddle2(saddleCandidate) and !dmt2PL[triangleId]){
+        if(isSaddle2(saddleCandidate) and dmt2Saddle2PL_[triangleId]==-1){
           saddleId=triangleId;
           ++numberOfSaddles;
         }
@@ -1462,8 +1510,8 @@ int DiscreteGradient::getRemovableSaddles2(const
 
       // only one DMT-2saddle in the star so this one is non-removable
       if(numberOfSaddles==1){
-        if(dmt2PL[saddleId]==false and pl2dmt_saddle[criticalPointId]==-1){
-          dmt2PL[saddleId]=true;
+        if(dmt2Saddle2PL_[saddleId]==-1 and pl2dmt_saddle[criticalPointId]==-1){
+          dmt2Saddle2PL_[saddleId]=criticalPointId;
           pl2dmt_saddle[criticalPointId]=saddleId;
           isRemovableSaddle[saddleId]=false;
         }
@@ -1588,7 +1636,6 @@ int DiscreteGradient::computeCoefficients(const bool isDense,
             }
           }
           if(!isIn)
-
             sparseCoefficients.push_back(std::make_pair(segmentId,segmentState));
         }
       }
@@ -1810,17 +1857,17 @@ isRemovableMaximum,
 }
 
 template <typename dataType>
-int DiscreteGradient::processSaddleMaximumConnections(const int
-                                                      iterationThreshold,
+int DiscreteGradient::processSaddleMaximumConnections(const int iterationThreshold,
                                                       const std::vector<char>& isPL,
                                                       const bool allowBoundary,
                                                       const bool allowBruteForce,
-                                                      std::set<std::pair<dataType,int>,SaddleMaximumVPathComparator<dataType>>& S,
+                                                      std::set<std::pair<dataType,int>,
+                                                       SaddleMaximumVPathComparator<dataType>>& S,
                                                       std::vector<simplexId_t>& pl2dmt_saddle,
                                                       std::vector<simplexId_t>& pl2dmt_maximum,
                                                       std::vector<Segment>& segments,
                                                       std::vector<VPath>& vpaths,
-                                                      std::vector<CriticalPoint>& criticalPoints) const{
+                                                      std::vector<CriticalPoint>& criticalPoints){
   Timer t;
 
   const dataType* const scalars=static_cast<dataType*>(inputScalarField_);
@@ -1877,10 +1924,8 @@ int DiscreteGradient::processSaddleMaximumConnections(const int
 
             simplexId_t saddleCandidateNumber=0;
             if(dimensionality_==2)
-
               saddleCandidateNumber=inputTriangulation_->getVertexEdgeNumber(pl_saddleId);
             else if(dimensionality_==3)
-
               saddleCandidateNumber=inputTriangulation_->getVertexTriangleNumber(pl_saddleId);
 
             for(simplexId_t j=0; j<saddleCandidateNumber; ++j){
@@ -1900,6 +1945,10 @@ int DiscreteGradient::processSaddleMaximumConnections(const int
 
             if(!numberOfRemainingSaddles){
               pl2dmt_saddle[vertexId]=dmt_saddleId;
+              if(dimensionality_==2)
+                dmt1Saddle2PL_[dmt_saddleId]=vertexId;
+              else if(dimensionality_==2)
+                dmt2Saddle2PL_[dmt_saddleId]=vertexId;
               vpath.invalidate();
               toRemoveSaddle=-1;
               break;
@@ -1961,6 +2010,7 @@ int DiscreteGradient::processSaddleMaximumConnections(const int
 
             if(!numberOfRemainingMaxima){
               pl2dmt_maximum[vertexId]=dmt_maxId;
+              dmtMax2PL_[dmt_maxId]=vertexId;
               vpath.invalidate();
               toRemoveMaximum=-1;
               break;
@@ -2089,7 +2139,6 @@ int DiscreteGradient::processSaddleMaximumConnections(const int
         S.erase(std::make_pair(newVPath.persistence_,newVPathId));
 
         // update persistence
-
         newVPath.persistence_=getPersistence<dataType>(newDestination.cell_,newSource.
           cell_, scalars);
 
@@ -2461,6 +2510,7 @@ int DiscreteGradient::processSaddleSaddleConnections1(const int
               if(numberOfRemainingSaddles1==0){
                 isRemovableSaddle1[dmt_saddle1Id]=false;
                 pl2dmt_saddle1[vertexId]=dmt_saddle1Id;
+                dmt1Saddle2PL_[dmt_saddle1Id]=vertexId;
                 vpath.invalidate();
                 break;
               }
@@ -2468,6 +2518,7 @@ int DiscreteGradient::processSaddleSaddleConnections1(const int
                 isRemovableSaddle1[dmt_saddle1Id]=false;
                 isRemovableSaddle1[savedId]=false;
                 pl2dmt_saddle1[vertexId]=savedId;
+                dmt1Saddle2PL_[savedId]=vertexId;
                 break;
               }
             }
@@ -2519,6 +2570,7 @@ int DiscreteGradient::processSaddleSaddleConnections1(const int
               if(numberOfRemainingSaddles2==0){
                 isRemovableSaddle2[dmt_saddle2Id]=false;
                 pl2dmt_saddle2[vertexId]=dmt_saddle2Id;
+                dmt2Saddle2PL_[dmt_saddle2Id]=vertexId;
                 vpath.invalidate();
                 break;
               }
@@ -2526,6 +2578,7 @@ int DiscreteGradient::processSaddleSaddleConnections1(const int
                 isRemovableSaddle2[dmt_saddle2Id]=false;
                 isRemovableSaddle2[savedId]=false;
                 pl2dmt_saddle2[vertexId]=savedId;
+                dmt2Saddle2PL_[savedId]=vertexId;
                 break;
               }
             }
@@ -3052,6 +3105,7 @@ int DiscreteGradient::processSaddleSaddleConnections2(const int
               if(numberOfRemainingSaddles1==0){
                 isRemovableSaddle1[dmt_saddle1Id]=false;
                 pl2dmt_saddle1[vertexId]=dmt_saddle1Id;
+                dmt1Saddle2PL_[dmt_saddle1Id]=vertexId;
                 vpath.invalidate();
                 break;
               }
@@ -3059,6 +3113,7 @@ int DiscreteGradient::processSaddleSaddleConnections2(const int
                 isRemovableSaddle1[dmt_saddle1Id]=false;
                 isRemovableSaddle1[savedId]=false;
                 pl2dmt_saddle1[vertexId]=savedId;
+                dmt1Saddle2PL_[savedId]=vertexId;
                 break;
               }
             }
@@ -3400,7 +3455,7 @@ int DiscreteGradient::filterSaddleConnectors(const bool allowBoundary){
       else if (upDegree == 1 and downDegree == 2)
         return 1;
     }
-      // local extremum
+    // local extremum
     else {
       if (upDegree)
         return 0;
@@ -3458,6 +3513,9 @@ int DiscreteGradient::reverseGradient(const
 
   std::vector<char> isPL;
   getCriticalPointMap(criticalPoints, isPL);
+
+  dmt1Saddle2PL_.resize(inputTriangulation_->getNumberOfEdges());
+  std::fill(dmt1Saddle2PL_.begin(), dmt1Saddle2PL_.end(), -1);
 
   if(ReverseSaddleMaximumConnection)
     simplifySaddleMaximumConnections<dataType>(criticalPoints, isPL,